requirements.txt
fastapi
uvicorn
python-decouple
sqlalchemy
psycopg2
pydantic
pydantic-settings
passlib
PyJWT
python-jose
cryptography